+++ packages/iov-faucets/types/iovfaucet.d.ts
@@ -1,0 +1,6 @@
+import { Address, TokenTicker } from "@iov/bcp-types";
+export declare class IovFaucet {
+    private readonly baseUrl;
+    constructor(baseUrl: string);
+    credit(address: Address, ticker: TokenTicker): Promise<void>;
+}