+++ packages/iov-faucets/types/utils.d.ts
@@ -1,0 +1,5 @@
+import { Address } from "@iov/bcp-types";
+/**
+ * Helper function for tests. Not exported in the package.
+ */
+export declare function randomBnsAddress(): Promise<Address>;