+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.8.0
 
+* @iov/faucets: Add `IovFaucet` to connect to the new faucet application
+
 Deprecations
 
 * @iov/faucets: `BovFaucet` is deprecated and will be removed in 0.9. Migrate to the `IovFaucet`.